# Conflicts:
#	config.gradle
#	lib-wallet-connect/src/main/java/com/violas/walletconnect/WCClient.kt
#	wallet-pay/src/main/java/com/violas/wallet/walletconnect/WalletConnect.kt

+++ lib-common/src/main/java/com/palliums/utils/DensityUtility.java
@@ -28,6 +28,7 @@ public class DensityUtility {
 
     /**
      * screenWidth
+     *
      * @param context
      * @return
      */
@@ -38,6 +39,7 @@ public class DensityUtility {
 
     /**
      * screenHeight
+     *
      * @param context
      * @return
      */
@@ -58,8 +60,8 @@ public class DensityUtility {
                 dpVal, context.getResources().getDisplayMetrics());
     }
 
-    public static int dp2px(Context context, float dpVal) {
-        return (int) TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP,
+    public static float dp2px(Context context, float dpVal) {
+        return TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP,
                 dpVal, context.getResources().getDisplayMetrics());
     }
